requirements.txt
django
mysqlclient
djangorestframework
PyJWT
python-dotenv
django-cors-headers
djangorestframework-simplejwt
pymongo